Добавлено логирование с сохранением логов в папку Logs. Добавлен Area для Admin. Во всех вьюшках админа убраны href=", заменены на asp-атрибуты
