+++ AirsoftShop/Areas/Admin/Contollers/UserController.cs
@@ -1,0 +1,13 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace AirsoftShop.Areas.Admin.Contollers
+{
+    [Area("Admin")]
+    public class UserController : Controller
+    {
+        public IActionResult Index()
+        {
+            return View();
+        }
+    }
+}
