+++ AirsoftShop/Areas/Admin/Views/Order/Detail.cshtml
@@ -100,7 +100,7 @@
                 </tr>
             </table>
 
-            <button class="save-button" type="submit" asp-controller="Admin" asp-action="UpdateOrderStatus" asp-route-orderId="@Model.Id">Обновить</button>
+            <button class="save-button" type="submit" asp-area="Admin" asp-controller="Order" asp-action="UpdateOrderStatus" asp-route-orderId="@Model.Id">Обновить</button>
         </form>
     </div>
 </div>
