+++ AirsoftShop/Areas/Admin/Views/Product/Add.cshtml
@@ -22,7 +22,6 @@
         display: block;
         margin-bottom: 8px;
         font-weight: bold;
-        color: #333;
     }
 
     input, textarea {
@@ -30,12 +29,13 @@
         padding: 10px;
         border: 1px solid #ced4da;
         border-radius: 4px;
-        font-size: 16px;
         box-sizing: border-box;
     }
 
-    textarea {
-        resize: vertical;
+    .form-buttons {
+        display: flex;
+        gap: 10px;
+        margin-top: 20px;
     }
 
     .btn {
@@ -43,8 +43,6 @@
         border: none;
         border-radius: 4px;
         cursor: pointer;
-        font-size: 16px;
-        transition: background-color 0.3s;
     }
 
     .btn-danger {
@@ -52,23 +50,13 @@
         color: white;
     }
 
-        .btn-danger:hover {
-            background-color: #c82333;
-        }
-
     .btn-secondary {
         background-color: #6c757d;
         color: white;
     }
 
-        .btn-secondary:hover {
-            background-color: #5a6268;
-        }
-
-    .form-buttons {
-        display: flex;
-        gap: 10px;
-        margin-top: 20px;
+    .text-danger {
+        color: #dc3545;
     }
 </style>
 
@@ -78,11 +66,9 @@
 
         <div class="col-9">
             <div class="admin-content">
-                <h3 class="text-center mb-4">Редактирование товара</h3>
+                <h3 class="text-center mb-4">Добавление нового товара</h3>
 
-                <form method="post" asp-controller="Admin" asp-action="EditProduct">
-                    <input type="hidden" asp-for="Id" />
-
+                <form method="post" asp-area="Admin" asp-controller="Product" asp-action="Add">
                     <div class="form-group">
                         <label asp-for="Name">Название товара</label>
                         <input asp-for="Name" class="form-control" required />
@@ -108,8 +94,8 @@
                     </div>
 
                     <div class="form-buttons">
-                        <button type="submit" class="btn btn-danger">Сохранить изменения</button>
-                        <a href="/admin/products" class="btn btn-secondary">Отмена</a>
+                        <button type="submit" class="btn btn-danger">Добавить товар</button>
+                        <a asp-area="Admin" asp-controller="Product" asp-action="Index" class="btn btn-secondary">Отмена</a>
                     </div>
                 </form>
             </div>
