+++ AirsoftShop/Areas/Admin/Views/Shared/AdminLeftMenu.cshtml
@@ -1,0 +1,14 @@
+﻿<div class="col-3">
+    <div class="row">
+        <a asp-area="Admin" asp-controller="Order" asp-action="Index">Заказы</a>
+    </div>
+    <div class="row">
+        <a asp-area="Admin" asp-controller="User" asp-action="Index">Пользователи</a>
+    </div>
+    <div class="row">
+        <a asp-area="Admin" asp-controller="Role" asp-action="Index">Роли</a>
+    </div>
+    <div class="row">
+        <a asp-area="Admin" asp-controller="Product" asp-action="Index">Товары</a>
+    </div>
+</div>
