+++ AirsoftShop/Areas/Admin/Views/User/Index.cshtml
@@ -1,0 +1,102 @@
+﻿
+
+<style>
+    .admin-panel {
+        font-family: Arial, sans-serif;
+        background-color: #f8f9fa;
+        padding: 20px;
+    }
+    
+    .admin-content {
+        background-color: white;
+        padding: 20px;
+        border-radius: 8px;
+        box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
+    }
+    
+    .user-table {
+        width: 100%;
+        border-collapse: collapse;
+    }
+    
+    .user-table th {
+        background-color: #f8f9fa;
+        padding: 15px;
+        text-align: left;
+        font-weight: bold;
+    }
+    
+    .user-table td {
+        padding: 15px;
+        border-bottom: 1px solid #ddd;
+    }
+    
+    .user-table tr:last-child td {
+        border-bottom: none;
+    }
+    
+    .btn {
+        padding: 6px 12px;
+        border: none;
+        border-radius: 4px;
+        cursor: pointer;
+        font-size: 14px;
+        transition: background-color 0.3s;
+        text-decoration: none;
+        display: inline-block;
+    }
+    
+    .btn-edit {
+        background-color: #ffc107;
+        color: #333;
+    }
+    
+    .btn-edit:hover {
+        background-color: #e0a800;
+    }
+    
+    .btn-delete {
+        background-color: #dc3545;
+        color: white;
+    }
+    
+    .btn-delete:hover {
+        background-color: #c82333;
+    }
+    
+    .btn-add {
+        background-color: #28a745;
+        color: white;
+        padding: 10px 20px;
+        font-size: 16px;
+        margin-bottom: 20px;
+    }
+    
+    .btn-add:hover {
+        background-color: #218838;
+    }
+</style>
+
+<div class="admin-panel">
+    <div class="row">
+        @await Html.PartialAsync("AdminLeftMenu")
+
+        <div class="col-9">
+            <div class="admin-content">
+                <h3 class="text-center mb-4">Список пользователей</h3>
+
+                <table class="user-table">
+                    <thead>
+                        <tr>
+                            #
+                        </tr>
+                    </thead>
+                    <tbody>
+                        <th>#</th>
+                    </tbody>
+                </table>
+                
+            </div>
+        </div>
+    </div>
+</div>
