+++ AirsoftShop/Areas/Admin/Views/_ViewImports.cshtml
@@ -1,0 +1,5 @@
+﻿@using AirsoftShop
+@using AirsoftShop.Models
+@using AirsoftShop.Helpers
+@using AirsoftShop.Areas.Admin.Models
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
