+++ AirsoftShop/Models/Login.cs
@@ -11,7 +11,7 @@ namespace AirsoftShop.Models
         [Required]
         [StringLength(25, MinimumLength = 6)]
         [RegularExpression(@"^(?=.*[A-Z])(?=.*\d)(?=.*[^\w\s]).{6,}$",
-            ErrorMessage = "Пароль должен содержать 1 заглавную букву, 1 цифру и 1 спецсимвол")]
+            ErrorMessage = "Пароль должен использовать латинские буквы, содержать 1 заглавную букву, 1 цифру и 1 спецсимвол")]
         public string Password { get; set; }
         public bool RememberMe { get; set; }
     }
