--- AirsoftShop/Views/Admin/Users.cshtml
@@ -1,7 +1,0 @@
-﻿<div class="row">
-@await Html.PartialAsync("AdminLeftMenu")
-
-    <div class="col-9">
-        <h3 d-flex justify-content-center>Список пользователй</h3>
-    </div>
-</div>
