--- AirsoftShop/Views/Shared/AdminLeftMenu.cshtml
@@ -1,14 +1,0 @@
-﻿<div class="col-3">
-    <div class="row">
-        <a href="/admin/orders">Заказы</a>
-    </div>
-    <div class="row">
-        <a href="/admin/users">Пользователи</a>
-    </div>
-    <div class="row">
-        <a href="/admin/roles">Роли</a>
-    </div>
-    <div class="row">
-        <a href="/admin/products">Товары</a>
-    </div>
-</div>
