+++ AirsoftShop/Views/Shared/Header.cshtml
@@ -35,16 +35,16 @@
                     </li>
 
                     <li class="nav-item ms-auto">
-                        <a class="nav-link text-dark custom-nav-link d-flex align-items-center" href="/cart/index">
+                        <a class="nav-link text-dark custom-nav-link d-flex align-items-center" asp-area="" asp-controller="Cart" asp-action="index">
                             Корзина
                             <span class="badge bg-primary ms-2" id="cartItemCount">@await Component.InvokeAsync("Cart")</span>
                         </a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link text-dark custom-nav-link" href="/account/login">Войти</a>
+                        <a class="nav-link text-dark custom-nav-link" asp-area="" asp-controller="account" asp-action="Login">Войти</a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link text-dark custom-nav-link" href="/Admin/orders">Администратор</a>
+                        <a class="nav-link text-dark custom-nav-link" asp-area="Admin" asp-controller="Order" asp-action="Index">Администратор</a>
                     </li>
                 </ul>
             </div>
