+++ AirsoftShop/appsettings.json
@@ -5,5 +5,30 @@
       "Microsoft.AspNetCore": "Warning"
     }
   },
-  "AllowedHosts": "*"
+  "AllowedHosts": "*",
+  "Serilog": {
+    "Using": [ "Serilog.Sinks.File", "Serilog.Sinks.Console" ],
+    "MinimumLevel": {
+      "Default": "Information",
+      "Override": {
+        "Microsoft": "Warning",
+        "System": "Warning"
+      }
+    },
+    "WriteTo": [
+      {
+        "Name": "Console"
+      },
+      {
+        "Name": "File",
+        "Args": {
+          "path": "C:\\Users\\Lutci\\source\\Repos\\AirsoftShop\\AirsoftShop\\Logs\\log-.json",
+          "rollOnFileSizeLimit": true,
+          "formatter": "Serilog.Formatting.Compact.CompactJsonFormatter,Serilog.Formatting.Compact",
+          "rollingInterval": "Day" // будет создаваться новый файл каждый денб
+        }
+      }
+    ],
+    "Enrich": [ "FromLogContext", "WithThreadId", "WithMachineName" ]
+  }
 }
